2014-02-27: 修改一些bug，优化一些项目（Inspecting code）

+++ App/src/main/java/com/df/app/entries/SerialNumber.java
@@ -101,7 +101,7 @@ public class SerialNumber {
             sn1 += snl;
             // sn2 contains the upper three digits of the serial number
             sn2 += snh;
-            return (((long)sn2) * 1000
+            return ((sn2) * 1000
                     + sn1);
         }
     }
